refactor(codegen): lift parse_ref / schema_base / enum_ref_target to shared refs
